% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farima.spectrum.R
\name{farima.spectrum}
\alias{farima.spectrum}
\title{Theoretical Spectral density of a FARIMA(p,d,q) process.}
\usage{
farima.spectrum(
  ar = 0,
  ma = 0,
  d = 0,
  var.noise = 1,
  n.freq = 512,
  frequency = c(0, 2 * pi)
)
}
\arguments{
\item{ar}{The coefficients of the autoregressive part of the process}

\item{ma}{The coefficients of the moving average part of the process}

\item{d}{# The long memory component, that should be between 0 and 0.5}

\item{var.noise}{# The variance of the process}

\item{n.freq}{# Length of the simulation}

\item{frequency}{# An interval of the domain in which the theoretical spectral density is evaluated}
}
\value{
Theoretical spectral density of a FARIMA process
}
\description{
It is based on an existing function from stats but modified to include long memory.
}
\examples{
farima.spectrum(ar = 0.5)
farima.spectrum(ar = c(0.5,-0.3), n.freq = 1024)
farima.spectrum(ar = 0.5, ma = -0.2, d = 0.3, n.freq = 8192)
}
