% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fexp.spectrum.R
\name{fexp.spectrum}
\alias{fexp.spectrum}
\title{Title}
\usage{
fexp.spectrum(
  ck = 0,
  d = 0,
  var.noise = 1,
  n.freq = 512,
  frequency = c(0, 2 * pi)
)
}
\arguments{
\item{frequency}{}
}
\description{
Title
}
