% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.farima.R
\name{sim.farima}
\alias{sim.farima}
\title{Function to simulate a FARIMA(p,d,q) process from its theoretical spectral density.}
\usage{
sim.farima(ar = 0, ma = 0, d = 0, T = 512)
}
\arguments{
\item{ar}{the autoregressive coefficients}

\item{ma}{the moving average coefficients}

\item{d}{the long memorry parameter}

\item{T}{the length of the time series to be simulated}
}
\value{
a time series of lenth T
}
\description{
Function to simulate a FARIMA(p,d,q) process from its theoretical spectral density.
}
\examples{
sim.farima(ar = 0.5)
sim.farima(ar = c(0.5,0.2),T = 128)
sim.farima(ar = 0.5, ma = -0.8, d =0.2, T = 8192)
}
