% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fexp.farima.R
\name{sim.fexp.farima}
\alias{sim.fexp.farima}
\title{Simulate the coresponding FEXP(p,d) process of a FARIMA(p,d,q) process}
\usage{
sim.fexp.farima(ar = 0, ma = 0, d = 0, T = 512, K = 2)
}
\arguments{
\item{K}{}
}
\description{
Simulate the coresponding FEXP(p,d) process of a FARIMA(p,d,q) process
}
