% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.fexp.R
\name{fit.fexp}
\alias{fit.fexp}
\title{Fit the FEXP(p,d) parameters using Whittle estimation}
\usage{
fit.fexp(y, p = 0, d = 0)
}
\arguments{
\item{d}{}
}
\description{
Fit the FEXP(p,d) parameters using Whittle estimation
}
